pyproject.toml
[project]
name = "xchange-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.13",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.2",
    "requests>=2.32.4",
]

[dependency-groups]
dev = [
    "ruff>=0.12.1",
]
